Balance and correct score calculating

+++ Assets/Scenes/Games/BaseGame/PlayingInfo.cs
@@ -2,7 +2,7 @@
 	public struct PlayingInfo {
 
 		public static string Name;
-		public static int Score;
+		public static float Score;
 		public static string GameMode;
 		public static float Time = 0;
